[fix] 기본값 설정을 위한 각 카테고리별 json 파일 추가
- 기본값 설정을 위한 각 카테고리별 json 파일 추가
- init.ts 파일에 지도 초기화 설정
