--- src/store/map/layers/landscape.ts
@@ -1,161 +1,0 @@
-export default [
-  {
-    id: 'landscape-humanmade',
-    type: 'fill',
-    source: 'polygon_source',
-    'source-layer': 'polygon',
-    layout: {
-      visibility: 'visible',
-    },
-    paint: {
-      'fill-color': 'grey',
-      'fill-opacity': 0.7,
-    },
-    filter: [
-      'in',
-      'type',
-      'civic',
-      'construction',
-      'gate',
-      'palace',
-      'parking',
-      'riverbank',
-      'station',
-      'steps',
-      'train_station',
-      'shrine',
-      'retaining_wall',
-      'smoking_area',
-      'city_gate',
-      'terrace',
-      'shelter',
-      'wall',
-      'fence',
-      'toll_booth',
-      'kerb',
-      'bench',
-      'footway',
-      'pedestrian',
-      'parking',
-      'fountain',
-      'gate',
-    ],
-  },
-  {
-    type: 'fill',
-    source: 'polygon_source',
-    'source-layer': 'polygon',
-    layout: {
-      visibility: 'visible',
-    },
-    paint: {
-      'fill-color': [
-        'interpolate',
-        ['linear'],
-        ['zoom'],
-        15,
-        'hsl(35, 11%, 86%)',
-        16,
-        'hsl(35, 7%, 85%)',
-      ],
-      'fill-opacity': ['interpolate', ['linear'], ['zoom'], 15, 0, 16, 1],
-      'fill-outline-color': 'hsl(35, 8%, 80%)',
-    },
-    filter: [
-      'in',
-      'type',
-      'apartments',
-      'bunker',
-      'cathedral',
-      'chapel',
-      'childcare',
-      'chungmoo art hall',
-      'church',
-      'courthouse',
-      'government',
-      'hospital',
-      'hotel',
-      'house',
-      'kindergarten',
-      'public',
-      'ranger_station',
-      'residential',
-      'roof',
-      'school',
-      'shed',
-      'shelter',
-      'university',
-      'warehouse',
-      'company',
-      'conference_centre',
-      'pavilion',
-      'swimming_pool',
-      'office',
-      'dormitory',
-      'public_bath',
-      'community_centre',
-      'theatre',
-      'fire_station',
-      'post_office',
-      'police',
-      'townhall',
-      'public',
-      'college',
-      'language_school',
-      'arts_centre',
-      'palace',
-      'toilets',
-      'social_facility',
-      'restaurant',
-      'cafe',
-      'marketplace',
-      'fast_food',
-      'bar',
-      'pub',
-      'casino',
-      'cinema',
-      '',
-    ],
-    id: 'landscape-building',
-  },
-  {
-    type: 'fill',
-    source: 'polygon_source',
-    'source-layer': 'polygon',
-    layout: {
-      visibility: 'visible',
-    },
-    paint: {
-      'fill-color': 'darkgreen',
-      'fill-opacity': 0.5,
-    },
-    filter: [
-      'match',
-      ['get', 'type'],
-      ['wood', 'scrub', 'bare_rock'],
-      true,
-      false,
-    ],
-    id: 'landscape-natural',
-  },
-  {
-    type: 'fill',
-    source: 'polygon_source',
-    'source-layer': 'polygon',
-    layout: {
-      visibility: 'visible',
-    },
-    paint: {
-      'fill-color': 'yellow',
-      'fill-opacity': 0.5,
-    },
-    filter: [
-      'match',
-      ['get', 'type'],
-      ['sand', 'heath', 'grassland'],
-      true,
-      false,
-    ],
-    id: 'landscape-landcover',
-  },
-];
