+++ src/utils/map-styling/landscape.ts
@@ -5,246 +5,99 @@ import {
   applyVisibility,
   WeightType,
   ColorType,
-  StyleTypes,
+  VisibilityType,
 } from '../applyStyle';
-import { StyleKeyType, ElementNameType } from '../../store/common/type';
+import {
+  StyleKeyType,
+  ElementNameType,
+  SubFeatureNameType,
+  objType,
+  SubElementNameType,
+} from '../../store/common/type';
+import layers from '../rendering-data/layers/landscape';
 
-const SECTION = 'section';
-const LABELTEXT = 'labelText';
-const LABELICON = 'labelIcon';
-
-type LandscapeSubFeature =
-  | 'all'
-  | 'human-made'
-  | 'building'
-  | 'natural'
-  | 'landcover'
-  | 'mountain';
-
-interface SubElementsType {
-  fill: string[];
-  stroke: string[];
-}
-
-interface ElementsType {
-  section: SubElementsType;
-  labelText: SubElementsType;
-  labelIcon: SubElementsType;
-}
-
-function makeSubElement(
-  fillLayers: string[],
-  strokeLayers: string[]
-): SubElementsType {
-  return { fill: fillLayers, stroke: strokeLayers };
-}
-
-const humanMadeSectionFill = ['landscape-human-made'];
-const humanMadeSectionStroke = ['pitch-outline'];
-
-const buildingSectionFill = ['landscape-building', 'building'];
-const buildingSectionStroke = ['building-outline'];
-const buildingLabel = ['building-number-label', 'poi-building-label'];
-
-const naturalSectionFill = ['landscape-natural', 'national-park'];
-const naturalLabel = ['natural-point-label'];
-
-const landCoverSectionFill = [
-  'land',
-  'landscape-landcover',
-  'landcover',
-  'landuse',
-  'land-structure-polygon',
-];
-const landCoverSectionStroke = ['land-structure-line'];
-
-const mountainCoverSectionFill = ['hillshade'];
-
-const layersByType: { [key in LandscapeSubFeature]: ElementsType } = {
-  'human-made': {
-    [SECTION]: makeSubElement(humanMadeSectionFill, humanMadeSectionStroke),
-    [LABELTEXT]: makeSubElement([], []),
-    [LABELICON]: makeSubElement([], []),
-  },
-  building: {
-    [SECTION]: makeSubElement(buildingSectionFill, buildingSectionStroke),
-    [LABELTEXT]: makeSubElement(buildingLabel, buildingLabel),
-    [LABELICON]: makeSubElement([], []),
-  },
-  natural: {
-    [SECTION]: makeSubElement(naturalSectionFill, []),
-    [LABELTEXT]: makeSubElement(naturalLabel, naturalLabel),
-    [LABELICON]: makeSubElement([], []),
-  },
-  landcover: {
-    [SECTION]: makeSubElement(landCoverSectionFill, landCoverSectionStroke),
-    [LABELTEXT]: makeSubElement([], []),
-    [LABELICON]: makeSubElement([], []),
-  },
-  mountain: {
-    [SECTION]: makeSubElement(mountainCoverSectionFill, []),
-    [LABELTEXT]: makeSubElement([], []),
-    [LABELICON]: makeSubElement([], []),
-  },
-  all: {
-    [SECTION]: makeSubElement(
-      [
-        ...humanMadeSectionFill,
-        ...buildingSectionFill,
-        ...naturalSectionFill,
-        ...landCoverSectionFill,
-        ...mountainCoverSectionFill,
-      ],
-      [
-        ...humanMadeSectionStroke,
-        ...buildingSectionStroke,
-        ...landCoverSectionStroke,
-      ]
-    ),
-    [LABELTEXT]: makeSubElement([], [...buildingLabel, ...naturalLabel]),
-    [LABELICON]: makeSubElement([], []),
-  },
+const init = {
+  section: { fill: [], stroke: [] },
+  labelText: { fill: [], stroke: [] },
+  labelIcon: [],
 };
 
-const VISIBLE = 1;
+const landscape = layers.reduce(
+  (prev, cur) => {
+    const [subFeature, element, subElement] = cur.id.split('-');
+    const list = prev;
+    if (!list.type[subFeature]) {
+      list.id[subFeature] = JSON.parse(JSON.stringify(init));
+      list.type[subFeature] = JSON.parse(JSON.stringify(init));
+    }
+    const { id, type } = cur;
+
+    if (element === ElementNameType.section) {
+      list.id[subFeature][element][subElement].push(id);
+      list.type[subFeature][element][subElement].push(type);
+
+      list.id[SubFeatureNameType.all][element][subElement].push(id);
+      list.type[SubFeatureNameType.all][element][subElement].push(type);
+    } else if (element === ElementNameType.labelText) {
+      list.id[subFeature][element][SubElementNameType.fill].push(id);
+      list.type[subFeature][element][SubElementNameType.fill].push(type);
+
+      list.id[subFeature][element][SubElementNameType.stroke].push(id);
+      list.type[subFeature][element][SubElementNameType.stroke].push(type);
+
+      list.id[SubFeatureNameType.all][element][SubElementNameType.fill].push(
+        id
+      );
+      list.type[SubFeatureNameType.all][element][SubElementNameType.fill].push(
+        type
+      );
+
+      list.id[SubFeatureNameType.all][element][SubElementNameType.stroke].push(
+        id
+      );
+      list.type[SubFeatureNameType.all][element][
+        SubElementNameType.stroke
+      ].push(type);
+    } else {
+      list.id[subFeature][element].push(id);
+      list.type[subFeature][element].push('icon');
+
+      list.id[SubFeatureNameType.all][element].push(id);
+      list.type[SubFeatureNameType.all][element].push('icon');
+    }
+    return list;
+  },
+  {
+    id: { all: JSON.parse(JSON.stringify(init)) },
+    type: { all: JSON.parse(JSON.stringify(init)) },
+  } as objType
+);
+
+type ColorVariableType =
+  | 'fill'
+  | 'line'
+  | 'text'
+  | 'background'
+  | 'icon'
+  | 'textHalo';
+type WeightVariableType = 'line' | 'textHalo';
+
+type StyleVariableType = ColorVariableType | WeightVariableType | 'symbol';
+interface getTypeProps {
+  element: ElementNameType;
+  subElement: SubElementNameType;
+  type: StyleVariableType;
+}
+
+const getType = ({ element, subElement, type }: getTypeProps) => {
+  if (element === ElementNameType.labelIcon) return 'icon';
+  if (type !== 'symbol') return type;
+
+  if (subElement === SubElementNameType.stroke) return 'textHalo';
+  return 'text';
+};
+
 const INVISIBLE = 0;
-
-const mappingDetailToFunc = {
-  labelText: {
-    fill: {
-      color: {
-        typeName: ColorType.text,
-        funcName: applyColor,
-      },
-      saturation: {
-        typeName: ColorType.text,
-        funcName: applyColor,
-      },
-      lightness: {
-        typeName: ColorType.text,
-        funcName: applyColor,
-      },
-      weight: {
-        typeName: null,
-        funcName: () => null,
-      },
-      visibility: {
-        typeName: 'what is my name?',
-        funcName: applyVisibility,
-      },
-      isChanged: {
-        typeName: null,
-        funcName: () => null,
-      },
-    },
-    stroke: {
-      color: {
-        typeName: ColorType.textHalo,
-        funcName: applyColor,
-      },
-      saturation: {
-        typeName: ColorType.textHalo,
-        funcName: applyColor,
-      },
-      lightness: {
-        typeName: ColorType.textHalo,
-        funcName: applyColor,
-      },
-      weight: {
-        typeName: WeightType.textHalo,
-        funcName: applyWeight,
-      },
-      visibility: {
-        typeName: WeightType.textHalo,
-        funcName: applyWeight,
-      },
-      isChanged: {
-        typeName: null,
-        funcName: () => null,
-      },
-    },
-  },
-  labelIcon: {
-    color: {
-      typeName: null,
-      funcName: () => null,
-    },
-    saturation: {
-      typeName: null,
-      funcName: () => null,
-    },
-    lightness: {
-      typeName: null,
-      funcName: () => null,
-    },
-    weight: {
-      typeName: null,
-      funcName: () => null,
-    },
-    visibility: {
-      typeName: ColorType.icon,
-      funcName: applyWeight,
-    },
-    isChanged: {
-      typeName: null,
-      funcName: () => null,
-    },
-  },
-  section: {
-    fill: {
-      color: {
-        typeName: ColorType.fill,
-        funcName: applyColor,
-      },
-      saturation: {
-        typeName: ColorType.fill,
-        funcName: applyColor,
-      },
-      lightness: {
-        typeName: ColorType.fill,
-        funcName: applyColor,
-      },
-      weight: {
-        typeName: null,
-        funcName: () => null,
-      },
-      visibility: {
-        typeName: ColorType.fill,
-        funcName: applyVisibility,
-      },
-      isChanged: {
-        typeName: null,
-        funcName: () => null,
-      },
-    },
-    stroke: {
-      color: {
-        typeName: ColorType.line,
-        funcName: applyColor,
-      },
-      saturation: {
-        typeName: ColorType.line,
-        funcName: applyColor,
-      },
-      lightness: {
-        typeName: ColorType.line,
-        funcName: applyColor,
-      },
-      weight: {
-        typeName: WeightType.line,
-        funcName: applyWeight,
-      },
-      visibility: {
-        typeName: WeightType.line,
-        funcName: applyVisibility,
-      },
-      isChanged: {
-        typeName: null,
-        funcName: () => null,
-      },
-    },
-  },
-};
 
 function landscapeStyling({
   map,
@@ -254,73 +107,66 @@ function landscapeStyling({
   key,
   style,
 }: stylingProps): void {
-  let type = null;
-  let func = null;
+  const idList = subElement
+    ? (landscape.id[subFeature][element][subElement] as string[])
+    : (landscape.id[subFeature][element] as string[]);
 
-  if (
-    element === ElementNameType.section ||
-    element === ElementNameType.labelText
-  ) {
-    const { typeName, funcName } = mappingDetailToFunc[element][subElement][
-      key as StyleKeyType
-    ];
-    type = typeName;
-    func = funcName;
-  } else {
-    const { typeName, funcName } = mappingDetailToFunc[element][
-      key as StyleKeyType
-    ];
-    type = typeName;
-    func = funcName;
-  }
+  const typeList = subElement
+    ? (landscape.type[subFeature][element][subElement] as StyleVariableType[])
+    : (landscape.type[subFeature][element] as StyleVariableType[]);
 
-  if (!type) {
+  if (key === StyleKeyType.weight) {
+    for (let i = 0; i < idList.length; i += 1) {
+      if (style.visibility !== VisibilityType.none) {
+        const type = getType({
+          element,
+          subElement,
+          type: typeList[i],
+        });
+        applyWeight({
+          map,
+          layerNames: [idList[i]],
+          type: WeightType[type as WeightVariableType] as WeightType,
+          weight: style[key],
+        });
+      }
+    }
     return;
   }
 
-  let layerNames =
-    layersByType[subFeature as LandscapeSubFeature][element][subElement];
-  if (
-    key === StyleKeyType.visibility &&
-    (type === ColorType.icon || type === WeightType.textHalo)
-  ) {
-    func({
-      map,
-      layerNames,
-      type,
-      weight: style.visibility === 'none' ? INVISIBLE : VISIBLE,
-    });
-    return;
-  }
   if (key === StyleKeyType.visibility) {
-    func({
-      map,
-      layerNames,
-      visibility: style.visibility,
-    });
+    for (let i = 0; i < idList.length; i += 1) {
+      const type = getType({ element, subElement, type: typeList[i] });
+      if (type === 'textHalo') {
+        applyWeight({
+          map,
+          layerNames: [idList[i]],
+          type: WeightType[type] as WeightType,
+          weight:
+            style.visibility === VisibilityType.none ? INVISIBLE : style.weight,
+        });
+      } else {
+        applyVisibility({
+          map,
+          layerNames: [idList[i]],
+          visibility: style[key],
+        });
+      }
+    }
+
+    return;
   }
 
-  if (
-    subElement === 'fill' &&
-    key !== StyleKeyType.visibility &&
-    layerNames.includes('land')
-  ) {
-    func({
+  for (let i = 0; i < idList.length; i += 1) {
+    const type = getType({ element, subElement, type: typeList[i] });
+    applyColor({
       map,
-      layerNames: ['land'],
-      type: ColorType.background,
+      layerNames: [idList[i]],
+      type: ColorType[type] as ColorType,
       color: style.color,
+      [key]: style[key],
     });
-    layerNames = layerNames.filter((item) => item !== 'land');
   }
-
-  func({
-    map,
-    layerNames,
-    type: type as StyleTypes,
-    color: style.color,
-    [key]: style[key as StyleKeyType],
-  });
 }
 
 export default landscapeStyling;
