--- src/utils/rendering-data/initLayerColor.ts
@@ -1,141 +1,0 @@
-interface InitLayerColor {
-  [name: string]: {
-    color: string;
-    type: string;
-  };
-}
-
-const initLayerColor: InitLayerColor = {
-  'poi-landmark-labelText-3': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-landmark-labelText-1': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-landmark-labelText-2': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-business-labelText-1': { color: 'hsl(22, 55%, 55%)', type: 'text' },
-  'poi-business-labelText-2': { color: 'hsl(22, 55%, 55%)', type: 'text' },
-  'poi-busniess-labelText-3': { color: 'hsl(22, 55%, 55%)', type: 'text' },
-  'poi-government-labelText-1': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-government-labelText-2': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-government-labelText-3': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-medical-labelText-1': { color: 'hsl(340, 39%, 42%)', type: 'text' },
-  'poi-medical-labelText-2': { color: 'hsl(340, 39%, 42%)', type: 'text' },
-  'poi-park-labelText-1': { color: 'hsl(100, 45%, 37%)', type: 'text' },
-  'poi-park-labelText-2': { color: 'hsl(100, 45%, 37%)', type: 'text' },
-  'poi-worship-labelText-1': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-worship-labelText-2': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-school-labelText-1': { color: 'hsl(51, 40%, 40%)', type: 'text' },
-  'poi-school-labelText-2': { color: 'hsl(51, 40%, 40%)', type: 'text' },
-  'poi-sport-labelText': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-etc-labelText-1': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-etc-labelText-2': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-etc-labelText-3': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'poi-building-label': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'admin-0-boundary': { color: 'hsl(230, 8%, 51%)', type: 'line' },
-  'admin-0-boundary-bg': { color: 'hsl(230, 8%, 51%)', type: 'line' },
-  'admin-0-boundary-disputed': { color: 'hsl(230, 8%, 51%)', type: 'line' },
-  'country-label': { color: 'hsl(0, 0%, 0%)', type: 'text' },
-  'admin-1-boundary-bg': { color: 'hsl(230, 14%, 77%)', type: 'line' },
-  'admin-1-boundary': { color: 'hsl(230, 14%, 77%)', type: 'line' },
-  'state-label': { color: 'hsl(0, 0%, 0%)', type: 'text' },
-  'settlement-label': { color: 'hsl(0, 0%, 0%)', type: 'text' },
-  'settlement-subdivision-label': { color: 'hsl(0, 0%, 0%)', type: 'text' },
-  water: { color: 'hsl(205, 87%, 76%)', type: 'fill' },
-  'water-polygon': { color: 'hsl(205, 87%, 76%)', type: 'fill' },
-  'water-line-label': { color: 'hsl(230, 48%, 44%)', type: 'text' },
-  'water-point-label': { color: 'hsl(230, 48%, 44%)', type: 'text' },
-  'waterway-label': { color: 'hsl(230, 48%, 44%)', type: 'text' },
-  //
-  'transit-airport-aeroway-section': {
-    color: 'hsl(234, 20%, 30%)',
-    type: 'fill',
-  },
-  'transit-airport-section': { color: 'hsl(234, 20%, 30%)', type: 'fill' },
-  'transit-airport-aeroway-line': { color: 'hsl(230, 23%, 82%)', type: 'line' },
-  'transit-airport-labelText': { color: 'hsl(230, 48%, 44%)', type: 'text' },
-  'transit-bus-labelText': { color: 'hsl(234, 20%, 30%)', type: 'text' },
-  'transit-rail-road-line': { color: 'hsl(234, 20%, 30%)', type: 'line' },
-  'transit-rail-line': { color: 'hsl(234, 20%, 30%)', type: 'line' },
-  'transit-subway-line': { color: 'hsl(192, 70%, 43%)', type: 'line' },
-  'road-arterial-section-fill-1': { color: 'hsl(0, 0%, 100%)', type: 'fill' },
-  'road-arterial-section-fill-2': { color: 'hsl(0, 0%, 100%)', type: 'line' },
-  'road-arterial-section-fill-3': { color: 'hsl(0, 0%, 100%)', type: 'line' },
-  'road-arterial-section-fill-4': { color: 'hsl(0, 0%, 100%)', type: 'line' },
-  'road-arterial-section-fill-5': { color: 'hsl(0, 0%, 100%)', type: 'line' },
-  'road-arterial-section-fill-6': { color: 'hsl(0, 0%, 100%)', type: 'line' },
-  'road-arterial-section-fill-7': { color: 'hsl(0, 0%, 100%)', type: 'line' },
-  'road-arterial-section-fill-8': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-arterial-section-fill-9': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-arterial-section-fill-10': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-arterial-section-fill-11': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-arterial-section-fill-12': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-arterial-section-fill-13': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-arterial-section-fill-14': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-arterial-section-fill-15': {
-    color: 'hsl(35, 10%, 83%)',
-    type: 'line',
-  },
-  'road-arterial-section-stroke-1': {
-    color: 'hsl(230, 24%, 87%)',
-    type: 'line',
-  },
-  'road-arterial-section-stroke-2': {
-    color: 'hsl(230, 24%, 87%)',
-    type: 'line',
-  },
-  'road-arterial-section-stroke-3': {
-    color: 'hsl(230, 24%, 87%)',
-    type: 'line',
-  },
-  'road-arterial-section-stroke-4': {
-    color: 'hsl(230, 24%, 87%)',
-    type: 'line',
-  },
-  'road-arterial-section-stroke-5': {
-    color: 'hsl(35, 10%, 83%)',
-    type: 'line',
-  },
-  'road-arterial-section-stroke-6': {
-    color: 'hsl(35, 10%, 83%)',
-    type: 'line',
-  },
-  'landscape-humanmade': { color: 'hsl(35, 11%, 86%)', type: 'fill' },
-  'road-arterial-labelText-1': { color: 'hsl(230, 48%, 44%)', type: 'text' },
-  'road-arterial-labelIcon-1': { color: 'hsl(0, 0%, 100%)', type: 'text' },
-  'road-arterial-labelIcon-2': { color: 'hsl(0, 0%, 100%)', type: 'text' },
-  'road-local-section-fill-1': { color: 'hsl(35, 14%, 93%)', type: 'fill' },
-  'road-local-section-fill-2': { color: 'hsl(230, 48%, 44%)', type: 'line' },
-  'road-local-section-fill-3': { color: 'hsl(35, 14%, 93%)', type: 'line' },
-  'road-local-section-fill-4': { color: 'hsl(35, 14%, 93%)', type: 'line' },
-  'road-local-section-fill-5': { color: 'hsl(35, 14%, 93%)', type: 'line' },
-  'road-local-section-stroke-1': { color: 'hsl(230, 24%, 87%)', type: 'line' },
-  'road-local-labelText-fill-1': { color: 'hsl(230, 48%, 44%)', type: 'text' },
-  'road-sidewalk-section-fill-1': { color: 'hsl(35, 10%, 83%)', type: 'fill' },
-  'road-sidewalk-section-fill-2': {
-    color: 'hsl(35, 10%, 83%)',
-    type: 'fill',
-  },
-  'road-sidewalk-section-fill-3': { color: 'hsl(35, 10%, 83%)', type: 'fill' },
-  'road-sidewalk-section-fill-4': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-sidewalk-section-fill-5': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-sidewalk-section-fill-6': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-sidewalk-section-fill-7': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-sidewalk-section-fill-8': { color: 'hsl(35, 10%, 83%)', type: 'line' },
-  'road-sidewalk-section-stroke-1': {
-    color: 'hsl(230, 24%, 87%)',
-    type: 'line',
-  },
-  'road-sidewalk-labelText-1': { color: 'hsl(230, 48%, 44%)', type: 'text' },
-  'pitch-outline': { color: 'hsl(75, 57%, 84%)', type: 'line' },
-  'landscape-building': { color: 'hsl(35, 11%, 86%)', type: 'fill' },
-  building: { color: 'hsl(35, 11%, 86%)', type: 'fill' },
-  'building-outline': { color: 'hsl(35, 6%, 79%)', type: 'line' },
-  'building-number-label': { color: 'hsl(35, 2%, 69%)', type: 'text' },
-  'landscape-natural': { color: 'hsl(75, 62%, 81%)', type: 'fill' },
-  'natural-point-label': { color: 'hsl(26, 25%, 32%)', type: 'text' },
-  'landscape-landcover': { color: 'hsl(75, 62%, 81%)', type: 'fill' },
-  landcover: { color: 'hsl(75, 62%, 81%)', type: 'fill' },
-  landuse: { color: 'hsl(75, 62%, 81%)', type: 'fill' },
-  'land-structure-polygon': { color: 'hsl(75, 62%, 81%)', type: 'fill' },
-  'land-structure-line': { color: 'hsl(35, 12%, 89%)', type: 'line' },
-  hillshade: { color: 'hsl(56, 59%, 22%)', type: 'fill' },
-};
-
-export default initLayerColor;
