+++ src/utils/rendering-data/layers/landscape.ts
@@ -1,76 +1,197 @@
-export enum LandScapeNameType {
-  all = 'all',
-  humanmade = 'humanmade',
-  building = 'building',
-  natural = 'natural',
-  landcover = 'landcover',
-  mountain = 'mountain',
-}
-
-const defaultStyle = {
-  all: {
-    section: {
-      fill: { color: 'transparent' },
-      stroke: { color: 'transparent' },
-    },
-    labelText: {
-      fill: { color: 'transparent' },
-      stroke: { color: 'transparent' },
-    },
-  },
-  humanmade: {
-    section: {
-      fill: { color: 'hsl(35, 11%, 86%)' },
-      stroke: { color: 'hsl(75, 57%, 84%)' },
-    },
-    labelText: {
-      fill: { color: 'transparent' },
-      stroke: { color: 'transparent' },
-    },
-  },
-  building: {
-    section: {
-      fill: { color: 'hsl(35, 11%, 86%)' },
-      stroke: { color: 'hsl(35, 6%, 79%)', weight: 1 },
-    },
-    labelText: {
-      fill: { color: 'hsl(35, 2%, 69%)' },
-      stroke: { color: 'hsl(0, 0%, 100%)', weight: 0.5 },
-    },
-  },
-  natural: {
-    section: {
-      fill: { color: 'hsl(75, 62%, 81%)' },
-      stroke: { color: 'transparent' },
-    },
-    labelText: {
-      fill: { color: 'hsl(26, 25%, 32%)' },
-      stroke: { color: 'hsl(0, 0%, 100%)', weight: 0.5 },
-    },
-  },
-  landcover: {
-    section: {
-      fill: { color: 'hsl(75, 62%, 81%)' },
-      stroke: { color: 'hsl(35, 12%, 89%)', weight: 1 },
-    },
-    labelText: {
-      fill: { color: 'transparent' },
-      stroke: { color: 'transparent' },
-    },
-  },
-  mountain: {
-    section: {
-      fill: { color: 'hsl(56, 59%, 22%)' },
-      stroke: { color: 'transparent' },
-    },
-    labelText: {
-      fill: { color: 'transparent' },
-      stroke: { color: 'transparent' },
-    },
-  },
-};
+import defaultStyle from '../defaultStyle';
 
 const layers = [
+  {
+    id: 'landcover-section-fill',
+    type: 'background',
+    metadata: {},
+    layout: {},
+    paint: {
+      'background-color': defaultStyle.landscape.landcover.section?.fill.color,
+    },
+  },
+  {
+    id: 'landcover-section-fill-2',
+    type: 'fill',
+    metadata: {},
+    source: 'composite',
+    'source-layer': 'landcover',
+    maxzoom: 7,
+    layout: {},
+    paint: {
+      'fill-color': defaultStyle.landscape.landcover.section?.fill.color,
+      'fill-opacity': [
+        'interpolate',
+        ['exponential', 1.5],
+        ['zoom'],
+        2,
+        0.3,
+        7,
+        0,
+      ],
+      'fill-antialias': false,
+    },
+  },
+  {
+    id: 'natural-section-fill-2',
+    type: 'fill',
+    metadata: {},
+    source: 'composite',
+    'source-layer': 'landuse_overlay',
+    minzoom: 5,
+    filter: ['==', ['get', 'class'], 'national_park'],
+    layout: {},
+    paint: {
+      'fill-color': defaultStyle.landscape.natural.section?.fill.color,
+      'fill-opacity': [
+        'interpolate',
+        ['linear'],
+        ['zoom'],
+        5,
+        0,
+        6,
+        0.5,
+        10,
+        0.5,
+      ],
+    },
+  },
+  {
+    id: 'landcover-section-fill-4',
+    type: 'fill',
+    metadata: {},
+    source: 'composite',
+    'source-layer': 'landuse',
+    minzoom: 5,
+    filter: [
+      'match',
+      ['get', 'class'],
+      [
+        'park',
+        'airport',
+        'cemetery',
+        'glacier',
+        'hospital',
+        'pitch',
+        'sand',
+        'school',
+      ],
+      true,
+      false,
+    ],
+    layout: {},
+    paint: {
+      'fill-color': defaultStyle.landscape.landcover.section?.fill.color,
+      'fill-opacity': [
+        'interpolate',
+        ['linear'],
+        ['zoom'],
+        5,
+        0,
+        6,
+        ['match', ['get', 'class'], 'glacier', 0.5, 1],
+      ],
+    },
+  },
+  {
+    id: 'landcover-section-fill-3',
+    type: 'fill',
+    source: 'polygon_source',
+    'source-layer': 'polygon',
+    layout: {
+      visibility: 'visible',
+    },
+    paint: {
+      'fill-color': defaultStyle.landscape.landcover.section?.fill.color,
+      'fill-opacity': 0.5,
+    },
+    filter: [
+      'match',
+      ['get', 'type'],
+      ['sand', 'heath', 'grassland'],
+      true,
+      false,
+    ],
+  },
+  {
+    id: 'humanmade-section-stroke',
+    type: 'line',
+    metadata: {},
+    source: 'composite',
+    'source-layer': 'landuse',
+    minzoom: 15,
+    filter: ['==', ['get', 'class'], 'pitch'],
+    layout: {},
+    paint: {
+      'line-color': defaultStyle?.landscape.humanmade.section?.stroke.color,
+    },
+  },
+  {
+    id: 'mountain-section-fill',
+    type: 'fill',
+    metadata: {},
+    source: 'composite',
+    'source-layer': 'hillshade',
+    maxzoom: 16,
+    layout: {},
+    paint: {
+      'fill-color': defaultStyle.landscape.mountain.section?.fill.color,
+      'fill-opacity': [
+        'interpolate',
+        ['linear'],
+        ['zoom'],
+        14,
+        ['match', ['get', 'level'], [67, 56], 0.06, [89, 78], 0.05, 0.12],
+        16,
+        0,
+      ],
+      'fill-antialias': false,
+    },
+  },
+  {
+    id: 'landcover-section-fill-5',
+    type: 'fill',
+    metadata: {},
+    source: 'composite',
+    'source-layer': 'structure',
+    minzoom: 13,
+    filter: [
+      'all',
+      ['==', ['geometry-type'], 'Polygon'],
+      ['==', ['get', 'class'], 'land'],
+    ],
+    layout: {},
+    paint: {
+      'fill-color': defaultStyle.landscape.landcover.section?.fill.color,
+    },
+  },
+  {
+    id: 'landcover-section-stroke',
+    type: 'line',
+    metadata: {},
+    source: 'composite',
+    'source-layer': 'structure',
+    minzoom: 13,
+    filter: [
+      'all',
+      ['==', ['geometry-type'], 'LineString'],
+      ['==', ['get', 'class'], 'land'],
+    ],
+    layout: { 'line-cap': 'round' },
+    paint: {
+      'line-width': defaultStyle.landscape.landcover.section?.stroke.weight,
+      // 'line-width': [
+      //   'interpolate',
+      //   ['exponential', 1.99],
+      //   ['zoom'],
+      //   14,
+      //   0.75,
+      //   20,
+      //   40,
+      // ],
+      'line-color': defaultStyle.landscape.landcover.section?.stroke.color,
+    },
+  },
   {
     id: 'humanmade-section-fill',
     type: 'fill',
@@ -80,7 +201,7 @@ const layers = [
       visibility: 'visible',
     },
     paint: {
-      'fill-color': defaultStyle.humanmade.section.fill.color,
+      'fill-color': defaultStyle.landscape.humanmade.section?.fill.color,
       'fill-opacity': 0.7,
     },
     filter: [
@@ -111,18 +232,36 @@ const layers = [
       'parking',
       'fountain',
       'gate',
+      '',
     ],
   },
   {
-    id: 'humanmade-section-stroke',
+    id: 'building-section-stroke',
     type: 'line',
     metadata: {},
     source: 'composite',
-    'source-layer': 'landuse',
+    'source-layer': 'building',
     minzoom: 15,
-    filter: ['==', ['get', 'class'], 'pitch'],
+    filter: [
+      'all',
+      ['!=', ['get', 'type'], 'building:part'],
+      ['==', ['get', 'underground'], 'false'],
+    ],
     layout: {},
-    paint: { 'line-color': defaultStyle.humanmade.section.stroke.color },
+    paint: {
+      'line-color': defaultStyle.landscape.building.section?.stroke.color,
+      'line-width': defaultStyle.landscape.building.section?.stroke.weight,
+      // 'line-width': [
+      //   'interpolate',
+      //   ['exponential', 1.5],
+      //   ['zoom'],
+      //   15,
+      //   0.75,
+      //   20,
+      //   3,
+      // ],
+      'line-opacity': ['interpolate', ['linear'], ['zoom'], 15, 0, 16, 1],
+    },
   },
   {
     id: 'building-section-fill',
@@ -133,7 +272,7 @@ const layers = [
       visibility: 'visible',
     },
     paint: {
-      'fill-color': defaultStyle.building.section.fill.color,
+      'fill-color': defaultStyle.landscape.building.section?.fill.color,
       'fill-opacity': ['interpolate', ['linear'], ['zoom'], 15, 0, 16, 1],
       'fill-outline-color': 'hsl(35, 8%, 80%)',
     },
@@ -212,40 +351,12 @@ const layers = [
         ['linear'],
         ['zoom'],
         15,
-        defaultStyle.building.section.fill.color,
+        defaultStyle.landscape.building.section?.fill.color,
         16,
         'hsl(35, 8%, 85%)',
       ],
       'fill-opacity': ['interpolate', ['linear'], ['zoom'], 15, 0, 16, 1],
       'fill-outline-color': 'hsl(35, 6%, 79%)',
-    },
-  },
-  {
-    id: 'building-section-stroke',
-    type: 'line',
-    metadata: {},
-    source: 'composite',
-    'source-layer': 'building',
-    minzoom: 15,
-    filter: [
-      'all',
-      ['!=', ['get', 'type'], 'building:part'],
-      ['==', ['get', 'underground'], 'false'],
-    ],
-    layout: {},
-    paint: {
-      'line-color': defaultStyle.building.section.stroke.color,
-      'line-width': defaultStyle.building.section.stroke.weight,
-      // 'line-width': [
-      //   'interpolate',
-      //   ['exponential', 1.5],
-      //   ['zoom'],
-      //   15,
-      //   0.75,
-      //   20,
-      //   3,
-      // ],
-      'line-opacity': ['interpolate', ['linear'], ['zoom'], 15, 0, 16, 1],
     },
   },
   {
@@ -263,9 +374,11 @@ const layers = [
       'text-size': 9.5,
     },
     paint: {
-      'text-color': defaultStyle.building.labelText.fill.color,
-      'text-halo-color': defaultStyle.building.labelText.stroke.color,
-      'text-halo-width': defaultStyle.building.labelText.stroke.weight,
+      'text-color': defaultStyle.landscape.building.labelText?.fill.color,
+      'text-halo-color':
+        defaultStyle.landscape.building.labelText?.stroke.color,
+      'text-halo-width':
+        defaultStyle.landscape.building.labelText?.stroke.weight,
     },
   },
   {
@@ -309,10 +422,12 @@ const layers = [
     },
     paint: {
       'icon-opacity': 1,
-      'text-halo-color': defaultStyle.building.labelText.stroke.color,
-      'text-halo-width': defaultStyle.building.labelText.stroke.weight,
+      'text-halo-color':
+        defaultStyle.landscape.building.labelText?.stroke.color,
+      'text-halo-width':
+        defaultStyle.landscape.building.labelText?.stroke.weight,
       'text-halo-blur': 0.5,
-      'text-color': defaultStyle.building.labelText.fill.color,
+      'text-color': defaultStyle.landscape.building.labelText?.fill.color,
     },
   },
   {
@@ -324,7 +439,7 @@ const layers = [
       visibility: 'visible',
     },
     paint: {
-      'fill-color': defaultStyle.natural.section.fill.color,
+      'fill-color': defaultStyle.landscape.natural.section?.fill.color,
       'fill-opacity': 0.5,
     },
     filter: [
@@ -334,30 +449,6 @@ const layers = [
       true,
       false,
     ],
-  },
-  {
-    id: 'natural-section-fill-2',
-    type: 'fill',
-    metadata: {},
-    source: 'composite',
-    'source-layer': 'landuse_overlay',
-    minzoom: 5,
-    filter: ['==', ['get', 'class'], 'national_park'],
-    layout: {},
-    paint: {
-      'fill-color': defaultStyle.natural.section.fill.color,
-      'fill-opacity': [
-        'interpolate',
-        ['linear'],
-        ['zoom'],
-        5,
-        0,
-        6,
-        0.5,
-        10,
-        0.5,
-      ],
-    },
   },
   {
     id: 'natural-labelText',
@@ -430,160 +521,11 @@ const layers = [
         17,
         ['step', ['get', 'sizerank'], 0, 13, 1],
       ],
-      'text-halo-color': defaultStyle.natural.labelText.stroke.color,
-      'text-halo-width': defaultStyle.natural.labelText.stroke.weight,
+      'text-halo-color': defaultStyle.landscape.natural.labelText?.stroke.color,
+      'text-halo-width':
+        defaultStyle.landscape.natural.labelText?.stroke.weight,
       'text-halo-blur': 0.5,
-      'text-color': defaultStyle.natural.labelText.fill.color,
-    },
-  },
-  {
-    id: 'landcover-section-fill',
-    type: 'background',
-    metadata: {},
-    layout: {},
-    paint: defaultStyle.landcover.section.fill.color,
-  },
-  {
-    id: 'landcover-section-fill-2',
-    type: 'fill',
-    metadata: {},
-    source: 'composite',
-    'source-layer': 'landcover',
-    maxzoom: 7,
-    layout: {},
-    paint: {
-      'fill-color': defaultStyle.landcover.section.fill.color,
-      'fill-opacity': [
-        'interpolate',
-        ['exponential', 1.5],
-        ['zoom'],
-        2,
-        0.3,
-        7,
-        0,
-      ],
-      'fill-antialias': false,
-    },
-  },
-  {
-    id: 'landcover-section-fill-3',
-    type: 'fill',
-    source: 'polygon_source',
-    'source-layer': 'polygon',
-    layout: {
-      visibility: 'visible',
-    },
-    paint: {
-      'fill-color': defaultStyle.landcover.section.fill.color,
-      'fill-opacity': 0.5,
-    },
-    filter: [
-      'match',
-      ['get', 'type'],
-      ['sand', 'heath', 'grassland'],
-      true,
-      false,
-    ],
-  },
-  {
-    id: 'landcover-section-fill-4',
-    type: 'fill',
-    metadata: {},
-    source: 'composite',
-    'source-layer': 'landuse',
-    minzoom: 5,
-    filter: [
-      'match',
-      ['get', 'class'],
-      [
-        'park',
-        'airport',
-        'cemetery',
-        'glacier',
-        'hospital',
-        'pitch',
-        'sand',
-        'school',
-      ],
-      true,
-      false,
-    ],
-    layout: {},
-    paint: {
-      'fill-color': defaultStyle.landcover.section.fill.color,
-      'fill-opacity': [
-        'interpolate',
-        ['linear'],
-        ['zoom'],
-        5,
-        0,
-        6,
-        ['match', ['get', 'class'], 'glacier', 0.5, 1],
-      ],
-    },
-  },
-  {
-    id: 'landcover-section-fill-5',
-    type: 'fill',
-    metadata: {},
-    source: 'composite',
-    'source-layer': 'structure',
-    minzoom: 13,
-    filter: [
-      'all',
-      ['==', ['geometry-type'], 'Polygon'],
-      ['==', ['get', 'class'], 'land'],
-    ],
-    layout: {},
-    paint: { 'fill-color': defaultStyle.landcover.section.fill.color },
-  },
-  {
-    id: 'landcover-section-stroke',
-    type: 'line',
-    metadata: {},
-    source: 'composite',
-    'source-layer': 'structure',
-    minzoom: 13,
-    filter: [
-      'all',
-      ['==', ['geometry-type'], 'LineString'],
-      ['==', ['get', 'class'], 'land'],
-    ],
-    layout: { 'line-cap': 'round' },
-    paint: {
-      'line-width': defaultStyle.landcover.section.stroke.weight,
-      // 'line-width': [
-      //   'interpolate',
-      //   ['exponential', 1.99],
-      //   ['zoom'],
-      //   14,
-      //   0.75,
-      //   20,
-      //   40,
-      // ],
-      'line-color': defaultStyle.landcover.section.stroke.color,
-    },
-  },
-  {
-    id: 'mountain-section-fill',
-    type: 'fill',
-    metadata: {},
-    source: 'composite',
-    'source-layer': 'hillshade',
-    maxzoom: 16,
-    layout: {},
-    paint: {
-      'fill-color': defaultStyle.mountain.section.fill.color,
-      'fill-opacity': [
-        'interpolate',
-        ['linear'],
-        ['zoom'],
-        14,
-        ['match', ['get', 'level'], [67, 56], 0.06, [89, 78], 0.05, 0.12],
-        16,
-        0,
-      ],
-      'fill-antialias': false,
+      'text-color': defaultStyle.landscape.natural.labelText?.fill.color,
     },
   },
 ];
